+++ tests/unit/demography/test_canopy.py
@@ -1,6 +1,7 @@
 """Testing the Canopy object."""
 
 import numpy as np
+import pytest
 
 
 def test_Canopy__init__():
@@ -31,7 +32,7 @@ def test_Canopy__init__():
     )
 
     canopy_gap_fraction = 0.05
-    canopy = Canopy(community, canopy_gap_fraction=canopy_gap_fraction)
+    canopy = Canopy(community, canopy_gap_fraction=canopy_gap_fraction, fit_ppa=True)
 
     # Simply check that the shape of the stem leaf area matrix is the right shape
     n_layers_from_crown_area = int(
@@ -46,4 +47,43 @@ def test_Canopy__init__():
             / community.cell_area
         )
     )
-    assert canopy.stem_leaf_area.shape == (n_layers_from_crown_area, canopy.n_cohorts)
+    assert canopy.layer_stem_leaf_area.shape == (
+        n_layers_from_crown_area,
+        canopy.n_cohorts,
+    )
+
+
+def test_solve_canopy_area_filling_height(fixture_community):
+    """Test solve_community_projected_canopy_area.
+
+    The logic of this test is that given the cumulative sum of the crown areas in the
+    fixture from tallest to shortest as the target, providing the z_max of each stem as
+    the height _should_ always return zero, as this is exactly the height at which that
+    cumulative area would close: crown 1 closes at z_max 1, crown 1 + 2 closes at z_max
+    2 and so on.
+    """
+
+    from pyrealm.demography.canopy import (
+        solve_canopy_area_filling_height,
+    )
+
+    for (
+        this_height,
+        this_target,
+    ) in zip(
+        np.flip(fixture_community.stem_allometry.crown_z_max),
+        np.cumsum(np.flip(fixture_community.stem_allometry.crown_area)),
+    ):
+        solved = solve_canopy_area_filling_height(
+            z=this_height,
+            stem_height=fixture_community.stem_allometry.stem_height,
+            crown_area=fixture_community.stem_allometry.crown_area,
+            n_individuals=fixture_community.cohort_data["n_individuals"],
+            m=fixture_community.stem_traits.m,
+            n=fixture_community.stem_traits.n,
+            q_m=fixture_community.stem_traits.q_m,
+            z_max=fixture_community.stem_allometry.crown_z_max,
+            target_area=this_target,
+        )
+
+    assert solved == pytest.approx(0)
